Prevent undefined index error on CLI
In some environments, php-cgi is the default binary. In this case, php_sapi_name() reports something like "cgi-fcgi" instead of the expected "cli". When raygun4php captures errors in this environment, it can't detect that PHP is running in a CLI environment and attempts to access some $_SERVER values that don't exist, causing PHP to throw an undefined index error. This patch prevents those errors.